v1.1

+++ Entities/Concrete/Car.cs
@@ -1,19 +1,17 @@
 ﻿using Core.Entities;
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
 
 namespace Entities.Concrete
 {
-    public class Car: IEntity
-    {        
+    public class Car : IEntity
+    {
         public int Id { get; set; }
         public int BrandId { get; set; }
         public int ColorId { get; set; }
         public int ModelYear { get; set; }
-        public decimal DailyPrice { get; set; }
+        public int DailyPrice { get; set; }
         public string Description { get; set; }
     }
 }
